Improve LSP messaging for ignored files
Currently, files that are ignored by Sorbet simply display a message of
the form:

```
[Error - 2:26:36 PM] Request textDocument/hover failed.
  Message: Did not find file at uri file:///path/to/file in textDocument/hover
  Code: -32602
```

which isn't very helpful when trying to determine the cause of the error
itself. Instead, I've updated the messaging for _ignored_ files
specifically along with a more appropriate LSPErrorCode.

```
[Error - 2:26:36 PM] Request textDocument/hover failed.
  Message: Ignored file at uri file:///path/to/file in textDocument/hover
  Code: -32803
```

this is more descriptive and provides the `LSPErrorCodes::RequestFailed`
error which is more descriptive of what occurred.