Invert compare_indent_level
It used to be the case that for something like

    def foo
      if
    end

if you were comparing the indent level of the `if` token (which is 2
spaces) and the `end` token (which is 0 spaces), that

    compare_indent_level(ifTok, endTok) == -1

which is weird because typically `(A <=> B) < 0` is true if `A < B` but
in this case we were effectively comparing `2 < 0` and saying that was
true.

It was just the case that I had also inverted the `>` in the one places
where this function is called, so it didn't matter.