Recover from mismatched def/class/end (#5383)
* Make stmts a delimited_node_list

* wip

* Factor out helper function

* defn_indent

* defs_indent

* class_indent

* Add some comments

* wip: Something is broken with newline_s

* truncateBeginBody -> truncateBodyStmt

* Use lineBreaks instead of newline_s

It looked like the do_nl action was happening too eagerly in certain
cases. For example:

    class
      Foo
    end

The lexer seemed to be running the `do_nl` action for the
`tCONSTANT(Foo)` token before calling `emit` for the `kCLASS` token. I
think that this is "working as intended" because the current use of the
`newline_s` rule seems to be that it's only guaranteed to have the right
thing in it when we're right about to emit a `tNL` token, and we don't
actually do that for the newline between `kCLASS` and `tCONSTANT` in
this case.

So what would happen instead was that the computed `lineStart` value for
the `kCLASS` token would take on a value **greater** than its own
`beginPos`.

* Fix class_weird_newline_indent test

* test for module_indent

* Add error assertions

* Some existing tests improve

* Move findLineBreaks into common

It seems benign enough to put there.

* Update some comments

* redundant include