Fix parser regression for `while ... do`
What was happening was that that having two rules using `kWHILE` forces
the parser to read a lookahead token to disambiguate which action to
take next.

- In the grammar before introducing the bug, it was unambiguous what to
  do next (immediately reduce the empty rule, running `cond.push` before
  lexing the next token).
- After the bug, it became ambiguous, so the grammar would read a token,
  and it would read `'('`. It just so happens that whenever we read one
  of those tokens, the **lexer** runs an additional `cond.push(false)`
  which it then pops when reading the matching `')'` (effectively saying
  that `do` means "do for block" inside `(...)`, even in a condition as
  long as there are parens around it.)

So the stack was supposed to look like

    ..., true, false

after reading a `(` token, but instead because the paren was getting
read as a lookahead token earlier than before, the stack looked like

    ..., false

after reading the `(` token, and then `true` was pushed later once the
parser decided to reduce the empty rule:

    ..., false, true

So upon reading the `)` token, which would pop the topmost boolean,
`cond.active()` would be `false` instead of `true`, which meant that
`"do"` lexed as `kDO` instead of `kDO_COND`.

- - - -

Some historical debugging context. Here's the good trace and the bad
trace:

```diff
❯ diff -u trace-good.txt trace-bad.txt
--- trace-good.txt      2022-03-07 13:32:18.662231915 -0800
+++ trace-bad.txt       2022-03-07 13:32:41.438250601 -0800
@@ -3,206 +3,182 @@
 Reading a token: Next token is token "while" ([0,5): )
 Shifting token "while" ([0,5): )
 Entering state 10
+Reading a token: Next token is token "(" ([6,7): )
 Reducing stack by rule 63 (line 866):
 driver.lex.cond.push(true)
 -> $$ = nterm $@4 ([0,5): )
 Stack now 10 0
-Entering state 262
-Reading a token: Next token is token "(" ([6,7): )
+Entering state 263
+Next token is token "(" ([6,7): )
 Shifting token "(" ([6,7): )
 Entering state 131
```

(computed by using `--trace-parser`, also I added some print statements
to show where the `cond.push(true)` happens)

We see that in the good trace (`-` lines), the `nterm $@4` reduction
happens before reading the `"("`, but in the bad trace (`+` lines), we
read a lookahead token before deciding to reduce one way or another.