Add isOpenInClient bit on core::File
This is currently unused, but we should be able to do this to power LSP
features for things that would otherwise be expensive to compute for
every file in a codebase. The use case that I have in mind right now is
visually distinguishing regions of a file that are `T.untyped`.