Entirely skip sorbet.run in publish on branch builds
This takes forever, because it has to clone sorbet.run, which has a ton
of binaries in it. Some numbers from a recent build:

- (98s) publish.sh job
  - (18s) cloning sorbet/sorbet itself, before running publish script
  - (53s) sorbet.run step of publish
  - (13s) vscode publish job

So half the time is just spent on sorbet.run nonsense.

On top of that, I can't remember it ever catching a legitimate bug. If
we ever need to make changes here, we can always tweak this code on a
branch build so that more of the steps run.