Updates signature and tests
The tests were not testing enough edgecases to cover different
permutations. As such, this commit adds those test cases and then
adjusts the signatures to work. The adjustments were to combine the
Enumerable generic and the solo type into a T.any set. It also moves the
explicit NilClass to the first signature and defines that it returns an
array of untyped.