Fix crash in YieldLoadParams (#5578)
* Add failing test

* Fix crash in YieldLoadParams

We have a helper function called getCallArguments that takes types like:

    T.proc.params(arg0: Integer, arg1: Integer).void

into types like:

    [Integer, Integer]

There is a special case that when youc all `getCallArguments` on
`T.untyped`, it will return `T.untyped`. Unfortunately, if you call it
on something that is typed but isn't a valid `getCallArguments` target
(like `BasicObject`) it would return `nullptr` instead of `T.untyped`
which would cause crashes downstream.