Adds type parameters to Kernel#Array (#5564)
* Adds type parameters to Kernel#Array

The types were being erased by the signature when sorbet is actually
smart enough to infer the correct types. This may cause new type checks
to fail but should be correct.

* Splats the input to the Array function

* Removes splat and defines an extra sig

This new signature will match on enumerable types and then map the
element of that enumerable to the type member of the T::Array.

* Updates signature and tests

The tests were not testing enough edgecases to cover different
permutations. As such, this commit adds those test cases and then
adjusts the signatures to work. The adjustments were to combine the
Enumerable generic and the solo type into a T.any set. It also moves the
explicit NilClass to the first signature and defines that it returns an
array of untyped.

* Adds final example from stripes tests that failed

* Adds nilable string example and updates sig

* update cli tests

Co-authored-by: Jake Zimmerman <zimmerman.jake@gmail.com>