Only run checkExpectations once (#5604)
checkExpectations will check **all** expectations. We were running it
twice for every exp kind except the `autocorrects` exp kind.

The diff is weird (it shows me as moving the "check warnings and errors"
code down below), but in my head this was moving the "Check
autocorrects" section up.

If we move things around like this, then by the time we get to
checkExpectations the first time, the autocorrects expectations will be
populated, so we don't have to run checkExpectations as second time.