Remove newlines and excess whitespace from autocorrects
When printing the messages into the terminal or into error messages in
VS Code, the extra whitespace is honestly more trouble than it's worth.

For a while I was pretty insistent that the autocorrect show you exactly
what it's going to do to your file, which mean showing the contents
unchanged. But honestly, people are just going to see that there's an
autocorrect and take it, and then undo the change if it turns out to
have not been a good one.