Add a section for subclassing T::Struct (#5734)
* Add a section for subclassing T::Struct

No where in this page does it explain that you can't derive a subclass from a parent that subclasses `T::Struct`. Nor is there any explanation of an alternative until after it fails and you look up the error code. This change links to the error code and providing an explanation that it can occur and how to overcome it.

* Update docs

Co-authored-by: Jake Zimmerman <zimmerman.jake@gmail.com>