srb init special case for ruby2_keywords in Ruby 3.1
Fixes #5741

This prints something different on Ruby 3.0 vs Ruby 3.1

    class A
      ruby2_keywords def foo(*args); end
    end
    p(A.instance_method(:foo).parameters)

Ruby 3.0

    [[:rest, :args]]

Ruby 3.1

    [[:rest, :args], [:keyrest, :**]]

I have no clue why this behavior changed, but it meant that we would
serialize a gem RBI that looked like

    class A
      def foo(*args, ****)
      end
    end

which is a syntax error.

I have decided not to write a test for this, because the test case would
have different behavior on Ruby 3.0 and Ruby 3.1, making it hard to fit
into our existing test snapshot testing framework.