Move ParsedArg to core
The ArgFlags were already in core (because they were shared with
ArgInfo, which is basically Symbol but for arguments).

The other two fields were also just a core::LocOffsets and a
core::LocalVariable.

I moved the ArgFlags to here so that I didn't have to depend on
`core/Types.h` from `core/ParsedArg.h`