Fix problem with aliased constants in VisibilityChecker (#5828)
* Add failing test

* Add another test

* Add another test

* findMember -> findMemberNoDealias in resolveConstant

* no-op: resolveConstants -> resolveResolutionItems

There were two methods on this class called resolveConstants, and they
just happened to have different types so overloading worked. But they
did completely different things. I noticed this while tryign to do jump
to definition and it presented me two jump targets in completely
separate spots.

* This test gains an error

It's maybe slightly unfortunate that this test gains this error, but I
don't see a way around it.

A constant with the name

    Opus::Devbox::Srv::Routes::UserFacing::Helpers::MAX_NUM_ADDITIONAL_DEVBOXES

does not exist. That is, there is nowhere in the code that it was
written:

    module Opus::Devbox::Srv::Routes::UserFacing::Helpers
      MAX_NUM_ADDITIONAL_DEVBOXES = ...
    end

or something similar. This constant _only_ resolves because there is a
class alias in the constant scope that makes it resolve:

    module Opus::Devbox::Srv::Routes::Helpers
      # This constant exists
      MAX_NUM_ADDITIONAL_DEVBOXES = ...
    end

    module Opus::Devbox::Srv::Routes::UserFacing
      Helpers = Opus::Devbox::Srv::Routes::Helpers
    end

This means that when we resolve the constant literal where the annotated
error in the diff, we have no choice but to resolve to the symbol that's
**not** in `UserFacing` (because no such constant with this name exists
in `UserFacing`).

That means that Sorbet then has to look up whether the resolved constant
is exported. In this case, it's not. Maybe that was intentional, because
the user was trying to essentially "re-export" private names to a public
namespace.

That's not going to work in Sorbet. There are two options:

1.  Declare the constant public in the first place, and alias it with a
    short name for private use, (instead of the other way around, like in
    this test).
2.  Re-export each individual constant, not a namespace of constants, so
    that Sorbet can track visibility for each constant, which would look
    like this:

    ```ruby
    module Opus::Devbox::Srv::Routes::UserFacing::Helpers
      MAX_NUM_ADDITIONAL_DEVBOXES =
        Opus::Devbox::Srv::Routes::Helpers::MAX_NUM_ADDITIONAL_DEVBOXES
      ...
    end
    ```

I'm personally satisfied with this. The error message could definitely
be better, but also this seems to be pretty rare? The changes in this PR
only cause three errors in Stripe's codebase. If this **really** becomes
a sticking point, we might even be able to walk the original
`ConstantLit`'s scope and see if any of the pieces of the scope were
exported, and then allow the reference.

* Add sigil to top of file