Better indentation-aware recovery for class & module (#5843)
Fixes #5841.

The comment in the parser is the best explanation of what was going on.

The key difference between class/module defs and method defs is that
method defs don't emit a tNL after the arguments end, while classe and
modules do. That means that classes and modules have their stmts list
constructed by starting from empty, reducing the `tNL`, and appending
the first element. Meanwhile methods start from the singleton
production, with no leading `tNL`.