methodArgumentHash: vector<uint32_t> -> uint32_t
Despite this being a vector, we were never using it as a vector really.
We were always just using it as a single hash value.

It will be convenient for #5808 if I can just call this method and get a
uint32_t, and since that already works for all the current use cases,
I've pulled it out as a pre-work change.