Remove GlobalState parameter
In the union for `LiteralType` storage, instead of storing a `uint32_t`
we simply store a `NameRef`. This means that in debug builds, we can
store extra information to track the GlobalState ID that created the
`NameRef`.

This also letes us remove the `GlobalState` parameter from `isString`,
`isSymbol`, `asString`, `asSymbol`, and `asName` (previously, the `gs`)
parameter was only being used for the sake of attributing a GlobalState
to the freshly minted NameRef (but there were no checks to make sure
that the `uint32_t` stored in the LiteralType union was actually created
by the `GlobalState` that had been passed into `asName`, so it was
possible to misuse).