Delete unused allowSameFromTo parameter
Aparently, I added this parameter back in `2317237b8db17a4bc`.

It looks like my reviewer stamped this without a second thought, but
this code was rather flawed. It introduced a boolean flag and threaded
it through ~every function in `NameSubstitution` just so that I could
silence an `ENFORCE` that would fail in a specific case of
`Substitute.cc`. That should never have been allowed to happen.

As of the previous commit, that one call site no longer sets this
boolean to true, so we can delete it.