Report histogram of how many files per fast path
The fast path could run on a single file or on tens of thousands of
files because of how our naive, name-based approach works.

We are going to be taking the fast path for more and more edits, which
will drive up the likelihood that we encounter worst-case edits that
typecheck many files on the fast path.

I want this metric so that we can get a sense of the magnitude.