pre-work: Add new UniqueNameKind::MangleRenameOverload
The way that overloads works fundamentally requires mangle renaming the
original method, so that there is still one method in the symbol table
with all the ArgInfo's, that we can selectively copy onto each (derived)
overload method.

That means that it definitely needs the logic that mangleRenameSymbol
does to rename things and erase things from the `members` vector.

But I want to do it in a way that's more predictable, so that it's
easier to map a name that was mangled in this way back to the original
method, as justified in the next commit.