Add mangleRenameForOverload method
Now, instead of calling mangleRenameSymbol in resolver, we call
mangleRenameForOverload. This lets us map back and forth between a
method that was mangled only for the purpose of overloading (ignoring
methods that have been mangle renamed in an attempt to recover from
method redefinitions).

This involves a slight trick in `lookupMethodSymbolWithHash`, because in
addition to walking up the chain of MangleRename unique names, it must
also look for a `MangleRenameOverload` at that step. In a picture:

Before:

    <U example>
        │
        ▼
    <M <U example> $1>
        │
        ▼
    <M <U example> $2>
        │
        ▼
        ⋮

Now:

    <U example>
        │
        ▼
    <M <U example> $1>  ─▶  <V <M <U example> $1> $1>
        ┌────────────────────────────┘
        ▼
    <M <U example> $2>  ─▶  <V <M <U example> $2> $1>
        ┌────────────────────────────┘
        ▼
        ⋮

With this in place, we can relax/rework some assumptions in
enterNewMethodOverload, namely: the method overload might already exist
(because we're on the fast path). The new `ENFORCE`'s all combine to
basically say that if there's already an method overload with the
current name, that it must match exactly the arity we're expecting.