Remove an error from the array_sum test
As far as I can tell, we only added this error so that we could avoid
triggering an ENFORCE when we made a prior Sorbet refactor. This code is
clearly not currently allowed in Sorbet, so it shouldn't matter if we
stop reporting an error here.

The only reason why we used to report an error here is because both
overloads and method redefinitions used `mangleRenameSymbol`, so when
attempting to run namer on `array_sum.rb`, `lookupMethodSymbolWithHash`
would return `<M <U sum> $1>`, the mangle rename created by resolver
when populating the payload for `array.rbi`.

After this change, `lookupMethodSymbolWithHash` looks at
`<U sum> (flags.isOverloaded=true)` but decides that it's overloaded, so
it should keep looking for `<V <U sum> $1>`, but it doesn't have the
same arity as the method def in `array_sum.rb` and so says "no existing
method with the expected arity" so it falls back to
`<U sum> (flags.isOverloaded=true)`. It happens that the first overload
of `Array#sum` has the right arity, so `paramMismatchErrors` reports no
errors, because there's actually no violation.

We can't do something like check whether the current method def is also
overloaded, because method defs can't be overloaded, only symbols. We
won't know whether this method def corresponds to an overload until
resolver looks at sigs.

So anyways, I couldn't think of a clean way to preserve the existing
error, but also I don't think it's worth preserving, so let's kill it.