Fix redefined & overloaded methods on fast_path (#5961)
* pre-work: Add new UniqueNameKind::MangleRenameOverload

The way that overloads works fundamentally requires mangle renaming the
original method, so that there is still one method in the symbol table
with all the ArgInfo's, that we can selectively copy onto each (derived)
overload method.

That means that it definitely needs the logic that mangleRenameSymbol
does to rename things and erase things from the `members` vector.

But I want to do it in a way that's more predictable, so that it's
easier to map a name that was mangled in this way back to the original
method, as justified in the next commit.

* Add mangleRenameForOverload method

Now, instead of calling mangleRenameSymbol in resolver, we call
mangleRenameForOverload. This lets us map back and forth between a
method that was mangled only for the purpose of overloading (ignoring
methods that have been mangle renamed in an attempt to recover from
method redefinitions).

This involves a slight trick in `lookupMethodSymbolWithHash`, because in
addition to walking up the chain of MangleRename unique names, it must
also look for a `MangleRenameOverload` at that step. In a picture:

Before:

    <U example>
        │
        ▼
    <M <U example> $1>
        │
        ▼
    <M <U example> $2>
        │
        ▼
        ⋮

Now:

    <U example>
        │
        ▼
    <M <U example> $1>  ─▶  <V <M <U example> $1> $1>
        ┌────────────────────────────┘
        ▼
    <M <U example> $2>  ─▶  <V <M <U example> $2> $1>
        ┌────────────────────────────┘
        ▼
        ⋮

With this in place, we can relax/rework some assumptions in
enterNewMethodOverload, namely: the method overload might already exist
(because we're on the fast path). The new `ENFORCE`'s all combine to
basically say that if there's already an method overload with the
current name, that it must match exactly the arity we're expecting.

* Two other edge cases

* Fix a bunch of tests

* Remove an error from the array_sum test

As far as I can tell, we only added this error so that we could avoid
triggering an ENFORCE when we made a prior Sorbet refactor. This code is
clearly not currently allowed in Sorbet, so it shouldn't matter if we
stop reporting an error here.

The only reason why we used to report an error here is because both
overloads and method redefinitions used `mangleRenameSymbol`, so when
attempting to run namer on `array_sum.rb`, `lookupMethodSymbolWithHash`
would return `<M <U sum> $1>`, the mangle rename created by resolver
when populating the payload for `array.rbi`.

After this change, `lookupMethodSymbolWithHash` looks at
`<U sum> (flags.isOverloaded=true)` but decides that it's overloaded, so
it should keep looking for `<V <U sum> $1>`, but it doesn't have the
same arity as the method def in `array_sum.rb` and so says "no existing
method with the expected arity" so it falls back to
`<U sum> (flags.isOverloaded=true)`. It happens that the first overload
of `Array#sum` has the right arity, so `paramMismatchErrors` reports no
errors, because there's actually no violation.

We can't do something like check whether the current method def is also
overloaded, because method defs can't be overloaded, only symbols. We
won't know whether this method def corresponds to an overload until
resolver looks at sigs.

So anyways, I couldn't think of a clean way to preserve the existing
error, but also I don't think it's worth preserving, so let's kill it.

* Add a comment about why we don't loop

* Update core/GlobalState.cc

Co-authored-by: Nathan Froyd <froydnj@gmail.com>

* Add an ENFORCE

Co-authored-by: Nathan Froyd <froydnj@gmail.com>