Use `if` instead of `||`
Using `||` in C++ on two integers results in a boolean condition. The
boolean is then converted to either `1` for `true` or `0` for false.

This means that it doesn't work like, say, JavaScript would, where it
would take the "truthy" (non-zero) one.

We pass this `rc` to `mdb_strerror`. If the `||` converts it to `1`,
then this looks like EPERM (which is errno `1` on Linux) instead of
something like `MDB_PAGE_FULL`.