Move DB version check into each cache flavor
This means that two flavors with the current (correct) DB version can
both live in the cache at the same time.

It also means that we can't get into a situation like this, which
happened in the real world before this PR:

    sorbet v1: before all this (fc3cf7a57^)
    sorbet v2: bad beta change (fc3cf7a57)
    sorbet v3: bad beta change + evict all flavors (dca4abd96)
    sorbet v4: evict all flavors (a485c5966)

steps:

1.  sorbet runs using v4, populates `default` flavor to shared cache
2.  an old branch runs, using v2 specifically, no cache eviction
    happens because it wasn't implemented yet. the `dsl-normalfastpath`
    flavor is created.
3.  build runs using v4. it sees that `default` flavor has the expected
    version (v4) and doesn't evict `dsl-normalfastpath`

This new change makes it so that a cache will only ever be as big all
the flavors for the current sorbet version require it to be, making
cache eviction simpler to reason about because now any version change
blows away all other versions in the cache (this was our mental model of
what happened before but it wasn't quite right in the presence of
multiple flavors in the database, as listed above).

It _does_ mean that if a given Sorbet version attempts to store multiple
flavors concurrently, the database now has to make sure there's enough
space for _all_ flavors to coexist, whereas the changes in #5974 would
have only let a given Sorbet version have one flavor in the database
(there could still be multiple flavors if they were for different
versions).