Add prodCounterSet
In case we attempt to commit the cache writes twice for one process, we
want to have the last one's size win. `prodCounterSet` makes this
happen.

I'm not actually sure this is required, and would be willing to replace
this with `prodCounterAdd` if it was always the case that we only called
it once.