Format compile_commands.json with `jq` (if available) (#6002)
A change a while back changed it so that the compile_commands.json file
is generated entirely on a single line. This makes it more difficult to
inspect manually, which I do from time to time to debug why it seems
like clangd isn't working correctly on a given file.