pre-work: Compute the owner centrally
This change makes it so that job.owner is always the actual owner
(sometimes a method, sometimes a class, depending on where the cast item
is), and we only widen it to `enclosingClass` right before we create a
`ctx` that we're going to pass into `getResultType`, which assumes that
`ctx.owner` is always a class, not a method.

This lets us use the method owner for future changes.