wip POC: Allow arbitrary type member applications
Right now you can only provide an explicit type application when calling
`.new` on a generic class. But every method that is typed as returning
`T.attached_class` could benefit from this syntax. This is a sketch of
an idea for how we could support it.

I haven't yet made up my mind on whether this is correct or even
desirable to implement this way. There are also some TODOs and code
cleanliness problems (duplication, variable names, etc.) that I have not
taken any time to address.