More descriptive exception messages
Sometimes we get a bug report that just says "should never happen."
Usually the only way to figure out what's up is to ask the user to
share a backtrace from a debugger like gdb or lldb so that there will
be line numbers, because there are too many "should never happen"
exceptions.

This makes it slightly easier to debug (though by no means perfect).