Fix crash in lub with tuples
Fixes #5286

The idea is that if we have to widen a proxy type to its underlying type
in order to attempt to collapse it with the other half of the `lub`, we
want to make that widening permanent.

If the code hadn't crashed via the ENFORCE, the revealed type in the
test would have been `[T.type_parameter(:U)] (1-tuple)`, which would
have been wrong, because the `when Array` branch doesn not necessarily
return a 1-tuple (that is what the ENFORCE correctly caught).

The crash looked like this:

    [2022-07-23 00:42:41.802] [fatalFallback] [error] Exception::raise(): core/types/subtyping.cc:49 enforced condition Types::isSubType(gs, t1, ret) has failed:
    TupleType {
      0 = SelfTypeParam(<C <U Object>>#<U foo>#<T <U U> $1>)
    }
    is not a super type of
    AppliedType {
          klass = <C <U Array>>
          targs = [
            <C <U Elem>> = T.untyped
          ]
        } & (AppliedType {
              klass = <C <U Array>>
              targs = [
                <C <U Elem>> = SelfTypeParam(<C <U Object>>#<U foo>#<T <U U> $1>)
              ]
            } | SelfTypeParam(<C <U Object>>#<U foo>#<T <U U> $1>))
    was lubbing with TupleType {
      0 = SelfTypeParam(<C <U Object>>#<U foo>#<T <U U> $1>)
    }