This edit is only to record notes to self:
All you're using the type constraint for is just to have an easy way to
replace a TypeVar with a "skolem," and in this case those skolems happen
to be SelfTypeParam.

(Dmitry's suggestion was to replace with skolems, so even though it
wasn't what I set out to try, I ended up following his advice
accidentally.)

The skolemization happens naturally by way of `Types::approximate`
(versus having to implement a whole new method purpose built for this.)

In this light, it isn't even necessary that the constraint "solve" or
even that it surve outside of `checkSubtype` to aggregate across all
argument (it could be allocated anew and destroyed right before the
`approximate` calls every time `checkSubtype` is called).

We can see this because the test actually passes when we use `isSubType`
instead of `isSubTypeUnderConstraint`. We still might want to thread the
constraint around everywhere so that we only ever have to allocate it
and do the `rememberIsSubType` calls once per method instead of once per
argument.

Note that this implementation still does not handle generic class type
members, but maybe the answer is just to sprinkle some calls to
`resultTypeAsSeenFrom` in `checkSubtype`? Haven't really thought that
one through yet.