Override checking for generic methods (#6086)
* wip: Thread TypeConstraint through validateCompatibleOverride

* Require that they have the same type param arity

* random: Allow SelfTypeParam::make_type on TypeArgumentRef

* wip: Bad attempt at something that didn't work

* wip: Second approach to generic method override

This one actually works on the example I was trying with?

    # typed: true

    module IFoo; end

    class Parent
      extend T::Sig
      sig do
        overridable
          .type_parameters(:U)
          .params(x: T.type_parameter(:U))
          .returns(T.type_parameter(:U))
      end
      def id(x); x; end
    end

    class ChildGood < Parent
      extend T::Sig
      sig do
        override
          .type_parameters(:V)
          .params(x: T.nilable(T.type_parameter(:V)))
          .returns(T.all(T.type_parameter(:V), IFoo))
      end
      def id(x)
        case x
        when nil then raise
        when IFoo then x
        else raise
        end
      end
    end

    class ChildBad < Parent
      extend T::Sig
      sig do
        override
          .type_parameters(:W)
          # .params(x: T.all(T.type_parameter(:W), IFoo))
          .params(x: T.all(T.type_parameter(:W), IFoo))
          .returns(T.nilable(T.type_parameter(:W)))
      end
      def id(x); x; end
    end

But it still doesn't handle type members, and I haven't bothered to test
it with proc types, and also there's the potentially glaring hole that
the `constr` doesn't actually solve at the bottom, which seems like
maybe a red flag but I haven't thought that through yet.

* wip: It appears that the lower bounds weren't required?

* Revert "wip: It appears that the lower bounds weren't required?"

This reverts commit 106cb7da5b615d1f7edba69f291e0e657e8b6d6c.

They actually are required when you throw in the changes I've made as a
part of #6076, which makes sense.

* Actually, don't even have to use the constraint for isSubType

This shows that we're basically only using the constraint as a way to
hijack approximate for skolemizing TypeVars to SelfTypeParams, and not
to accumulate new constraints during the process of going across all the
args.

* wip: Add a test

* This edit is only to record notes to self:

All you're using the type constraint for is just to have an easy way to
replace a TypeVar with a "skolem," and in this case those skolems happen
to be SelfTypeParam.

(Dmitry's suggestion was to replace with skolems, so even though it
wasn't what I set out to try, I ended up following his advice
accidentally.)

The skolemization happens naturally by way of `Types::approximate`
(versus having to implement a whole new method purpose built for this.)

In this light, it isn't even necessary that the constraint "solve" or
even that it surve outside of `checkSubtype` to aggregate across all
argument (it could be allocated anew and destroyed right before the
`approximate` calls every time `checkSubtype` is called).

We can see this because the test actually passes when we use `isSubType`
instead of `isSubTypeUnderConstraint`. We still might want to thread the
constraint around everywhere so that we only ever have to allocate it
and do the `rememberIsSubType` calls once per method instead of once per
argument.

Note that this implementation still does not handle generic class type
members, but maybe the answer is just to sprinkle some calls to
`resultTypeAsSeenFrom` in `checkSubtype`? Haven't really thought that
one through yet.

* Move git log into comment

* wip: Do it for type members too

* Add test for class generics

* another todo

* Don't check if no sig

* Don't show approximated types

Reasoning is that

- The code is terrible, and refactoring this would be mildly painful.
- On second thought, it's actually better this way, because it will show
  the type failure in terms closer to what the user wrote in the file.

* Add some more docs to override-checking

* Add some more tests

* Add another test

* Update core/Types.h

Co-authored-by: Nathan Froyd <froydnj@gmail.com>

* Add sanity check

Co-authored-by: Nathan Froyd <froydnj@gmail.com>