Allow `abstract` methods in `# typed: strong` files
The change as implemented forces the user to choose either `# compiled:
true` or `# typed: strong` for a file containing an `abstract` method.

This might also contribute a slight performance improvement, by skipping
`CFG` construction and inference for abstract methods.

This comes at the cost of no longer type checking default arguments
against their provided types in abstract methods, but arguably this is
fine, as the default values for an abstract method's parameters are
never actually used at runtime.