LSP ErrorReporter: Do not report extraneous empty error lists to the client.
Sorbet only needs to send empty error lists to the client when it has previously sent errors for a file that now has no errors.

Before this change, when the client has `lspErrorCap` errors, the ErrorReporter would report an empty error list for
any file with errors that exceed the cap -- even if Sorbet had never sent errors for that file before.
This can freeze VS Code (DoS due to many messages) if you make a catastrophic typo that causes many files to error.

This PR changes `clientErrorCount` to count the errors _reported_ to the client (as it should), and not the total number of errors in the last epoch.
Similarly, `FileErrorStatus.errorCount` now tracks the errors _reported_ to the client, and not the errors in the last epoch.

Now, we can avoid sending unneeded empty lists when `errorCount == 0` and `errorsToReport == 0`.