Skip already-added files in fast path decision
This _should_ make the list of files free of duplicates by construction.

It also makes it easier for us to have the `gs->getFiles()` loop short
circuit once it adds too many files (deferring to a future change
because that might come after we move this logic from LSPTypechecker.cc
to LSPIndexer.cc).