Use UnorderedSet instead of epoch
epoch was wrong because I assumed that every `core::File` within
`updates.updatedFiles` would have its `epoch` set to the epoch of the
change.

That's not the case, the epoch is decided and never mutated at the time
the `DidChange` for that file happens. Which means that if the changes
come in as individual DidChanges and then they're processed on the same
fast path edit (because they were lumped together), one of them will
have an old-looking epoch (that is less than `updates.epoch`).

I had been trying to avoid using an UnorderedSet for this to avoid
having to allocate extra things but I think we can offset this easily:

- The `updates` variable has the indexed trees for the files from the
  update. If it's easy for us to distinguish between which files were in
  the edit and not, we can get rid of jvilk's TODO to skip indexing for
  already-indexed files.

  This would be nice because it would mean we don't e.g., parse a file
  **4 times** to service one completion request (e.g. if there was
  a syntax error we will index once to find an empty file, again in
  indentation-aware mode, and then twice more here in runFastPath).

- In addition to being able to short circuit from the "extraFiles" loop,
  skipping the set intersection, we should also be able to start
  completely short circuiting this loop and jump back to the slow path
  if there are too many extra files. That should be deferred to a future
  change.