Never suggest `# typed: ignore` (#6164)
* Never suggest `# typed: ignore`

We shouldn't suggest `# typed: ignore` because it is too common for some
probem in a generated RBI file to cause a problem that actually does
need to be fixed, and not by ignoring the RBI file. Ignoring the file
has bad consequences, like introducing more errors in other files,
causing those files to be ignored, etc.

This causes a huge number of problems and the user is never told when it
has happened. This problem especially confuses people attempting to
adopt Sorbet for the first time, and they have no clue that it's
happening.

* Update exp file