Merge evicted files before running canTakeFastPath
Some files might only be present in one of the two sets of files. In the
implementation of canTakeFastPath, we assume that if the evictedFiles
mapping is not empty, but that is not true unless we merge things
beforehand like this.

You should probably be able to write a multithreaded protocol test for
this?

It would have to be for something like three edits:

1.  There is a slow path running
2.  There is a fast path after that (else we would have cancelled the
    running slow path)
3.  There is a slow path after that (else we would not get back to
    this call to mergeEvictedFiles)
    - merging edits (2) + (3) can take the fast path given that (1) is
      still running (else our fastPathFilesToTypecheck helper would
      not have been called)
    - edit (2) must mention a file that (3) didn't change

Having written that out, it sounds like the test could possibly be:

1.  Slow path edit in `foo.rb`
2.  Edit in method body in `bar.rb`
3.  Undo edit in `foo.rb`