Don't track blame if method doesn't exist (#6190)
Fixes #4818

There are some places where we might have untyped because a method
didn't exist, like this:

    # typed: strict
    module DatabaseInit
      class CreateUsers
        def do
          existing_roles = Set.new(rows.map { |row| row['rolname'] })
        end
      end
    end

In this example, the `rows` method does not exist, which means that when
we set the resultType for the blockResultType and blockPreType, we were
previously setting it to a symbol that didn't exist.

I thought about fixing the one-off problems with blocks in
environment.cc, but then I realized that it seemed like basically any
place where the method didn't exist, the behavior would be to simply
create an untracked untyped at that call site. So I figured I'd just
push that logic into the implementation of `Types::untyped`.