Fix infinite loop in type_member with self (#6184)
* Fix infinite loop in type_member with self

Fixes #3613

These two things should behave identically:

    class A
      extend T::Generic
      Elem1 = type_member {{upper: A}}
      Elem2 = type_member {{upper: self}}
    end

We correctly report the error for `Elem1` by tracking a list of
dependencies which must have been resolved before we allow resolving the
`ResolveAssignItem` that will enter the bounds for the type member.

However, we were forgetting to add the dependencies when creating a job
that mentions `self` instead of a constant literal.

This manifested as an infinite loop when trying to use the upper bound
to do something recursive. The first place that happened was when
computing hashes for LSP, because LSP will hash the type members of a
class, which will hash their upper bounds, which will hash the class,
which will hash the type members of the class (which is a cycle).

(The fact that `self` is allowed here at all is maybe an accident of
history. I don't quite know why it behaves the way it does, but `self`
is basically allowed in any type syntax position where the syntactic
name of the enclosing class would also be allowed. It's a weird rule but
it's what we have.)

* Fix tests