Redefine methods with correct visibility upfront (#6215)
Fixes #5786

Sorbet runtime updates the visibility of the methods it is replacing
but sometimes that happens too late. For example, if user code has a
`method_added` hook and checks the visibility of the method that's
added in the hook, it would always see `public` when Sorbet
replaces a `private` or `protected` method. This was happening because
the setting of the visibility was happening after the call to
`define_method` and by then it was too late.

Instead of setting the visibility after defining the method, this PR
encapsulates a method definition operation inside a method named
`def_with_visibility` on the `T::Private::ClassUtils` module, which
does a `module_exec` on the passed in module so that it can create a
`public`/`protected`/`private` section and only then define the method.
This ensures that the method is defined with the expected visibility
from the beginning.