Add type_alias test that passes but shouldn't
This is the same problem as captured in #5940, but using `type_alias`
instead of `alias_method`.

For `alias_method`, we made the choice that this was rare enough that it
didn't matter, and that we should ship first and maybe figure out a way
to fix it later. I'm not sure whether we want to also do that for
`T.type_alias`, or whether we want to block on being able to fix this.