Fix bug with scope lookup
Apart from the renames (I struggled more than I would like to admit
keeping this code straight while debugging, because it is highly
recursive), the main change is the `lookupClassSymbol`.

That call makes sure that it finds the right symbol kind in the presence
of previous mangle renames. Note that there are only two places where we
call `ensureScopeIsClass`, and the other one (in `insertStaticField`)
also uses `lookupStaticFieldSymbol` instead of `findMember`.

Without this change, what would happen is that given a test case like
this:

    module A::B::C
      def foo; end
    end

    A::B = A::D

On the incremental path, the `findMember` call would look up `A::B` as a
static field. And now that the static field properly stores the alias to
`StubModule`, class `C` would get entered on `StubModule` instead of on
the existing `A::B::C` class symbol.

(This test exists as `nested_module_then_nonexistent_name.rb`.)